Move more entries from the deploy script to the
deploy module.
